Test repo, success save Employees

+++ src/main/java/com/baga/promon/usermanagement/domain/Employee.java
@@ -1,5 +1,6 @@
 package com.baga.promon.usermanagement.domain;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
-public record Employee(Long id, String address, String name, OffsetDateTime joinDate) {}
+public record Employee(BigDecimal id, String address, String name, OffsetDateTime joinDate) {}
